Fixed time measurements. Work measurement is correct. Still major BUGS in span measurement.

+++ include/FunctionStack.h
@@ -46,31 +46,31 @@ struct function_frame_t {
 	*    @var local_work
 	*    @brief Local work of the function.
 	*/
-	double local_work;
+	uint64_t local_work;
 
 	/**
 	*    @var local_lock_span
 	*    @brief The local lock span of the function.
 	*/
-	double local_lock_span;
+	uint64_t local_lock_span;
 
 	/**
 	*    @var running_work
 	*    @brief The running work of the function's call site. 
 	*/
-	double running_work;
+	uint64_t running_work;
 
 	/**
 	*    @var running_work
 	*    @brief The running lock span of the function's call site. 
 	*/
-	double running_lock_span;
+	uint64_t running_lock_span;
 
 	/**
 	*    @var running_span
 	*    @brief The running span of the function's call site. 
 	*/
-	double running_span;
+	uint64_t running_span;
 
 	function_frame_t() : call_site(0)
 	{}
